requirements.txt
Pillow
scipy==1.0.0
scikit-learn
scikit-image
dill
numpy
opencv-python
tensorflow-gpu==1.14.0
jupyterlab
imageio
noise
matplotlib
imutils
pyyaml
joblib
torch
torchvision
addict
click
dominate
dill
requests
pydensecrf
tqdm
ipywidgets
